Update mimemagic dependency
`mimemagic` in older versions required projects to be under the General
Public License. It came to light recently and it was a mistake of the
mimemagic's package creator.

We do not want the project to be kept under GPL license, so we upgrade
the dependency so as to allow ourselves to keep it under MIT one.

Resolves https://github.com/michudzik/dissertation/issues/2